+++ epam-promo/components/layout/IconContainer.tsx
@@ -7,7 +7,7 @@ type IconContainerColors = 'blue' | 'green' | 'amber' | 'red' | 'cyan' | 'orange
 
 interface IconContainerMods {
     /**
-     *  IconContainer color.
+     *  Defines component color.
      *  @deprecated Property color is deprecated and will be removed in future release. Please make icon color configuration by yourself, e.g. via cx or style prop.
      *  @default 'gray60'
      */
@@ -21,7 +21,8 @@ function applyIconContainerMods(mods: IconContainerMods) {
     ];
 }
 
-export type IconContainerProps = uuiComponents.ControlIconProps & IconContainerMods;
+/** Represents the properties of a IconContainer component. */
+export interface IconContainerProps extends uuiComponents.ControlIconProps, IconContainerMods {}
 
 export const IconContainer = createSkinComponent<uuiComponents.ControlIconProps, IconContainerProps>(
     uui.IconContainer,