+++ loveship/components/overlays/NotificationCard.tsx
@@ -1,18 +1,25 @@
-import { withMods } from '@epam/uui-core';
+import { createSkinComponent, devLogger } from '@epam/uui-core';
 import * as uui from '@epam/uui';
 import { EpamPrimaryColor } from '../types';
 
-export interface NotificationCardMods {
-    color: EpamPrimaryColor | 'night600' | uui.NotificationCardProps['color'];
+interface NotificationCardMods {
+    /**
+     * Defines component color.
+     */
+    color?: EpamPrimaryColor | 'night600' | uui.NotificationCardProps['color'];
 }
 
-export type NotificationCardProps = Omit<uui.NotificationCardProps, 'color'> & NotificationCardMods;
+/** Represents the properties of the NotificationCard component. */
+export interface NotificationCardProps extends uui.NotificationCardCoreProps, NotificationCardMods {}
 
-export const NotificationCard = withMods<Omit<uui.NotificationCardProps, 'color'>, NotificationCardMods>(
+export const NotificationCard = createSkinComponent<uui.NotificationCardProps, NotificationCardProps>(
     uui.NotificationCard,
-    () => [],
-    (props) => ({
-        ...props,
-        color: props.color ?? null,
-    }),
+    (props) => {
+        if (__DEV__) {
+            if (props.color === 'night600') {
+                devLogger.warn('(NotificationCard) The night600 value of color is deprecated and will be removed in future release. Use color sky instead or consult with your design team.');
+            }
+        }
+        return { ...props, color: props.color };
+    },
 );