+++ uui/components/layout/FlexItems/Panel.tsx
@@ -3,11 +3,21 @@ import { VPanel } from '@epam/uui-components';
 import css from './Panel.module.scss';
 
 interface PanelMods {
+    /*
+    * Pass true to show a shadow.
+    */
     shadow?: boolean;
+    /*
+    * Defines component margin.
+    */
     margin?: '24';
+    /*
+    * Defines component background.
+    */
     background?: 'surface-main';
 }
 
+/** Represents the properties of the Panel component. */
 export type PanelProps = VPanelProps & PanelMods;
 
 export const Panel = withMods<VPanelProps, PanelMods>(VPanel, (props) => [