+++ uui/components/layout/LabeledInput.tsx
@@ -5,17 +5,18 @@ import { Tooltip } from '../overlays/Tooltip';
 import { systemIcons } from '../../icons/icons';
 import css from './LabeledInput.module.scss';
 
-const defaultSize = '36';
+const DEFAULT_SIZE = '36';
 
-export interface LabeledInputMods extends types.SizeMod {}
+interface LabeledInputMods extends types.SizeMod {}
 
+/** Represents the properties of the LabeledInput component. */
 export type LabeledInputProps = uuiComponents.LabeledInputProps & LabeledInputMods;
 
 function applyLabeledInputMods(mods: LabeledInputMods) {
-    return [css.root, css['size-' + (mods.size || defaultSize)]];
+    return [css.root, css['size-' + (mods.size || DEFAULT_SIZE)]];
 }
 
 export const LabeledInput = withMods<uuiComponents.LabeledInputProps, LabeledInputMods>(uuiComponents.LabeledInput, applyLabeledInputMods, (props) => ({
     Tooltip,
-    infoIcon: systemIcons[props.size || defaultSize].help,
+    infoIcon: systemIcons[props.size || DEFAULT_SIZE].help,
 }));